+++ src/constants/solidity.test.ts
@@ -7,7 +7,7 @@ test('exports solidity constants', () => {
     {
       "panicReasons": {
         "1": "An \`assert\` condition failed.",
-        "17": "Arithmic operation resulted in underflow or overflow.",
+        "17": "Arithmetic operation resulted in underflow or overflow.",
         "18": "Division or modulo by zero (e.g. \`5 / 0\` or \`23 % 0\`).",
         "33": "Attempted to convert to an invalid type.",
         "34": "Attempted to access a storage byte array that is incorrectly encoded.",